Add files via upload
This is code for pentagon plot which is a straightforward method of graphic representation of five bands in two-dimensional space. Spectral bands are assigned to vertexes of the pentagon. Each multispectral sample is represented by a set of vectors pointing towards the respective vertices with their starting point in the centre of the pentagon. The magnitude of the vectors is given by reflectance values. The position of the sample in the plot is determined by vector addition. This is followed by scaling in the radial direction to achieve a better distribution of points in the plot area.